devtools/shared/security/tests/chrome/test_websocket-transport.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test the WebSocket debugger transport</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<script>
"use strict";

window.onload = function() {
  const {require} = ChromeUtils.import("resource://devtools/shared/Loader.jsm");
  const Services = require("Services");
  const {DebuggerClient} = require("devtools/shared/client/debugger-client");
  const {DebuggerServer} = require("devtools/server/main");
  const { SocketListener } = require("devtools/shared/security/socket");

  Services.prefs.setBoolPref("devtools.debugger.remote-enabled", true);
  Services.prefs.setBoolPref("devtools.debugger.prompt-connection", false);

  SimpleTest.registerCleanupFunction(() => {
    Services.prefs.clearUserPref("devtools.debugger.remote-enabled");
    Services.prefs.clearUserPref("devtools.debugger.prompt-connection");
  });

  add_task(async function() {
    DebuggerServer.init();
    DebuggerServer.registerAllActors();

    is(DebuggerServer.listeningSockets, 0, "0 listening sockets");

    const socketOptions = {
      portOrPath: -1,
      webSocket: true,
    };
    const listener = new SocketListener(DebuggerServer, socketOptions);
    ok(listener, "Socket listener created");
    await listener.open();
    is(DebuggerServer.listeningSockets, 1, "1 listening socket");

    const transport = await DebuggerClient.socketConnect({
      host: "127.0.0.1",
      port: listener.port,
      webSocket: true,
    });
    ok(transport, "Client transport created");

    const client = new DebuggerClient(transport);
    const onUnexpectedClose = () => {
      ok(false, "Closed unexpectedly");
    };
    client.addListener("closed", onUnexpectedClose);

    await client.connect();

    // Send a message the server that will echo back
    const message = "message";
    const reply = await client.mainRoot.echo({ message });
    is(reply.message, message, "Echo message matches");

    client.removeListener("closed", onUnexpectedClose);
    transport.close();
    listener.close();
    is(DebuggerServer.listeningSockets, 0, "0 listening sockets");

    DebuggerServer.destroy();
  });
};
</script>
</body>
</html>
